requirements.txt
# Computer Vision
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.3.0

# Hugging Face & Transformers
transformers==4.41.2
accelerate==0.31.0
sentence-transformers==2.7.0
datasets==2.19.1
tokenizers==0.19.1

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilities
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
pyyaml==6.0.1
tqdm==4.66.4
requests==2.32.3
colorama==0.4.6

# FAISS - CPU version
faiss-cpu==1.8.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Audio processing
librosa==0.10.2.post1
soundfile==0.12.1

# Logging
loguru==0.7.2

# Testing
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0

# Face Recognition
facenet-pytorch==2.5.3

# Audio Analysis
librosa==0.10.2.post1